about to start cleaning up note_instructions_to_timing_changes

+++ src/functions/create_long_note.py
@@ -4,10 +4,10 @@ from src.classes import LongNote
 def create_long_note(note_list: list[str], key_count: int) -> LongNote:
     column = int(note_list[0])
     y = int(note_list[1])
-    time = int(note_list[2])
+    time = float(note_list[2])
     note_type = int(note_list[3])
     hit_sound = int(note_list[4])
-    end_time = int(note_list[5])
+    end_time = float(note_list[5])
     hit_sample = note_list[6]
 
     new_note = LongNote(
